Move getParamsForUpdateOrInsert into controller
