+++ libraries/classes/Controllers/Table/ReplaceController.php
@@ -25,6 +25,7 @@ use PhpMyAdmin\Transformations;
 use PhpMyAdmin\Util;
 
 use function __;
+use function array_keys;
 use function array_values;
 use function class_exists;
 use function implode;
@@ -104,7 +105,7 @@ final class ReplaceController extends AbstractController
         /**
          * Prepares the update/insert of a row
          */
-        [$loopArray, $usingKey, $isInsert] = $this->insertEdit->getParamsForUpdateOrInsert();
+        [$loopArray, $usingKey, $isInsert] = $this->getParamsForUpdateOrInsert();
 
         $isInsertignore = isset($_POST['submit_type']) && $_POST['submit_type'] === 'insertignore';
 
@@ -505,5 +506,37 @@ final class ReplaceController extends AbstractController
          */
         /** @psalm-suppress UnresolvableInclude */
         require ROOT_PATH . Core::securePath($gotoInclude);
+    }
+
+    /**
+     * Prepares the update/insert of a row
+     *
+     * @return mixed[] $loop_array, $using_key, $is_insert
+     * @psalm-return array{array, bool, bool}
+     */
+    private function getParamsForUpdateOrInsert(): array
+    {
+        if (isset($_POST['where_clause'])) {
+            // we were editing something => use the WHERE clause
+            $loopArray = is_array($_POST['where_clause'])
+                ? $_POST['where_clause']
+                : [$_POST['where_clause']];
+            $usingKey = true;
+            $isInsert = isset($_POST['submit_type'])
+                && ($_POST['submit_type'] === 'insert'
+                    || $_POST['submit_type'] === 'showinsert'
+                    || $_POST['submit_type'] === 'insertignore');
+        } else {
+            // new row => use indexes
+            $loopArray = [];
+            if (! empty($_POST['fields'])) {
+                $loopArray = array_keys($_POST['fields']['multi_edit']);
+            }
+
+            $usingKey = false;
+            $isInsert = true;
+        }
+
+        return [$loopArray, $usingKey, $isInsert];
     }
 }
